Move ASPayload class to seperate file.
This to support PSR-4 Auto loading as when we need to use/import the class.
